+++ lib/less/rails/bootstrap/engine.rb
@@ -5,7 +5,7 @@ module Less
     module Bootstrap
       class Engine < ::Rails::Engine
         
-        initializer 'less-rails-bootstrap.setup', :after => 'less-rails.after.load_config_initializers', :group => :all do |app|
+        initializer 'less-rails-bootstrap.setup', :after => 'less-rails.before.load_config_initializers', :group => :all do |app|
           app.config.less.paths << File.join(config.root, 'app', 'frameworks')
         end